+++ resources/leiningen/new/luminus/compojure/src/home.clj
@@ -5,26 +5,24 @@
             [ring.util.http-response :as response]
             [clojure.java.io :as io]))
 <% if cljs  %>
-(defn home-page []
-  (layout/render "home.html"))
+(defn home-page [request]
+  (layout/render request "home.html"))
 
 (defroutes home-routes
-  (GET "/" []
-       (home-page))<% if graphql %>
-  (GET "/graphiql" [] (layout/render "graphiql.html"))<% endif %>
+  (GET "/" request (home-page request))<% if graphql %>
+  (GET "/graphiql" request (layout/render request "graphiql.html"))<% endif %>
   (GET "/docs" []
        (-> (response/ok (-> "docs/docs.md" io/resource slurp))
            (response/header "Content-Type" "text/plain; charset=utf-8"))))
 <% else %>
-(defn home-page []
-  (layout/render
-    "home.html" {:docs (-> "docs/docs.md" io/resource slurp)}))
+(defn home-page [request]
+  (layout/render request "home.html" {:docs (-> "docs/docs.md" io/resource slurp)}))
 
-(defn about-page []
-  (layout/render "about.html"))
+(defn about-page [request]
+  (layout/render request "about.html"))
 
 (defroutes home-routes
-  (GET "/" [] (home-page))<% if graphql %>
-  (GET "/graphiql" [] (layout/render "graphiql.html"))<% endif %>
-  (GET "/about" [] (about-page)))
+  (GET "/" request (home-page request))<% if graphql %>
+  (GET "/graphiql" request (layout/render request "graphiql.html"))<% endif %>
+  (GET "/about" request (about-page request)))
 <% endif %>