+++ resources/leiningen/new/luminus/core/src/layout.clj
@@ -6,14 +6,13 @@
             [ring.util.anti-forgery :refer [anti-forgery-field]]
             [ring.middleware.anti-forgery :refer [*anti-forgery-token*]]))
 
-<% if servlet %>(declare ^:dynamic *app-context*)<% endif %>
 (parser/set-resource-path!  (clojure.java.io/resource "html"))
 (parser/add-tag! :csrf-field (fn [_ _] (anti-forgery-field)))
 (filters/add-filter! :markdown (fn [content] [:safe (md-to-html-string content)]))
 
 (defn render
   "renders the HTML template located relative to resources/html"
-  [template & [params]]
+  [request template & [params]]
   (content-type
     (ok
       (parser/render-file
@@ -21,8 +20,8 @@
         (assoc params
           :page template
           :csrf-token *anti-forgery-token*<% if servlet %>
-          :selmer/context *app-context*
-          :servlet-context *app-context*<% endif %>)))
+          :selmer/context (-> request :session :app-context)
+          :servlet-context (-> request :session :app-context)<% endif %>)))
     "text/html; charset=utf-8"))
 
 (defn error-page